requirements.txt
selenium~=4.11.2
python-dotenv~=1.0.0
Selenium-Screenshot~=2.1.0
